+++ scout/parse/variant/variant.py
@@ -255,17 +255,17 @@ def set_severity_predictions(parsed_variant: dict, variant: dict, parsed_transcr
     parsed_variant["spidex"] = call_safe(float, variant.INFO.get("SPIDEX"))
 
     if len(parsed_transcripts) > 0:
-        parsed_variant["revel_score"] = parsed_transcripts[0].get(
-            "revel_rankscore"
-        )  # This is actually the value of REVEL_rankscore
-        parsed_variant["revel"] = get_highest_revel_score(parsed_transcripts)
+        parsed_variant["revel_score"] = get_highest_revel_score(
+            parsed_transcripts=parsed_transcripts, key="revel_rankscore"
+        )  # will be displayed as "REVEL rank score"
+        parsed_variant["revel"] = get_highest_revel_score(
+            parsed_transcripts=parsed_transcripts, key="revel_raw_score"
+        )  # will be displayed as "REVEL score"
 
 
-def get_highest_revel_score(parsed_transcripts: List[dict]) -> Optional[float]:
-    """Retrieve the highest REVEL_score value from parsed variant transcripts."""
-    tx_revel_scores: List(float) = [
-        tx.get("revel_raw_score") for tx in parsed_transcripts if tx.get("revel_raw_score") != None
-    ]
+def get_highest_revel_score(parsed_transcripts: List[dict], key: str) -> Optional[float]:
+    """Retrieve the highest REVEL score values from parsed variant transcripts."""
+    tx_revel_scores: List(float) = [tx.get(key) for tx in parsed_transcripts if tx.get(key) != None]
     if tx_revel_scores:
         return max(tx_revel_scores)